@@ -2,9 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { BlankComponent } from './layouts/blank/blank.component';
 import { FullComponent } from './layouts/full/full.component';
-import { AuthGuard } from './guards/auth.guard';
-// import { IntroGuard } from './guards/intro.guard';
-// import { AutoLoginGuard } from './guards/auto-login.guard';
+import { AuthGuard } from '../guards/auth.guard';
 
 const routes: Routes = [
   {
@@ -21,18 +19,18 @@ const routes: Routes = [
         loadChildren: () => import('./pages/pages.module').then((m) => m.PagesModule),
         canLoad: [AuthGuard] // Check if we should show the introduction or fo
       },
-      // {
-      //   path: 'ui-components',
-      //   loadChildren: () =>
-      //     import('./pages/ui-components/ui-components.module').then(
-      //       (m) => m.UicomponentsModule
-      //     ),
-      // },
-      // {
-      //   path: 'extra',
-      //   loadChildren: () =>
-      //     import('./pages/extra/extra.module').then((m) => m.ExtraModule),
-      // },
+    //   // {
+    //   //   path: 'ui-components',
+    //   //   loadChildren: () =>
+    //   //     import('./pages/ui-components/ui-components.module').then(
+    //   //       (m) => m.UicomponentsModule
+    //   //     ),
+    //   // },
+    //   // {
+    //   //   path: 'extra',
+    //   //   loadChildren: () =>
+    //   //     import('./pages/extra/extra.module').then((m) => m.ExtraModule),
+    //   // },
     ],
   },
   {
