@@ -4,23 +4,28 @@ import { CommonModule } from '@angular/common';
 import { PagesRoutes } from './pages.routing.module';
 import { MaterialModule } from '../material.module';
 import { FormsModule } from '@angular/forms';
-import { NgApexchartsModule } from 'ng-apexcharts';
+// import { NgApexchartsModule } from 'ng-apexcharts';
 // icons
 import { TablerIconsModule } from 'angular-tabler-icons';
 import * as TablerIcons from 'angular-tabler-icons/icons';
-import { AppDashboardComponent } from './dashboard/dashboard.component';
+// import { AppDashboardComponent } from './dashboard/dashboard.component';
 import { FullCalendarModule } from '@fullcalendar/angular';
+import {CalendarComponent} from "./calendar/calendar.component";
+import {DashboardComponent} from "./dashboard/dashboard.component";
 
 @NgModule({
-  declarations: [AppDashboardComponent],
+  declarations: [
+    CalendarComponent,
+    DashboardComponent
+  ],
   imports: [
     CommonModule,
     MaterialModule,
     FormsModule,
-    NgApexchartsModule,
+    // NgApexchartsModule,
     RouterModule.forChild(PagesRoutes),
     TablerIconsModule.pick(TablerIcons),
-    FullCalendarModule
+    FullCalendarModule,
   ],
   exports: [TablerIconsModule],
 })
