@@ -1,13 +1,14 @@
 import { Routes } from '@angular/router';
-import { AppDashboardComponent } from './dashboard/dashboard.component';
+// import { AppDashboardComponent } from './dashboard/dashboard.component';
 import { CalendarComponent } from './calendar/calendar.component';
+import {DashboardComponent} from "./dashboard/dashboard.component";
 
 export const PagesRoutes: Routes = [
   {
     path: 'dashboard',
-    component: AppDashboardComponent,
+    component: DashboardComponent,
     data: {
-      title: 'Starter Page',
+      title: 'Dashboard',
     },
   },
   {
